+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## [TBD]
+### Changed
 * Refactored project to remove checkstyle and findbugs warnings, which does not impact functionality.
 * Upgraded `hotels-oss-parent` to 2.3.5 (was 2.3.3).
+### Fixed
+* Fixed compatibility layer exception handling.
 
 ## [3.1.1] - 2018-10-30
 ### Fixed